stx-update: API Ref Doc Content Enabled
API Reference content has been added for stx-update:

  - Add link and pages for API Ref and main page.

The project-specific docs are published independently so we need
to maintain the links into those subtrees, for the api-ref (/api-ref/stx-*).

This change was introduced and then reverted due to this abandoned
commit:
  https://review.openstack.org/#/c/605278/
Reason was a confussion during the split of old API into the new
API scheme.

Depends-On: https://review.openstack.org/#/c/634994/
Change-Id: I3a4bd5e98cf7077cf4c3abe037d1a0f6d8c69fcf
Signed-off-by: Abraham Arce <abraham.arce.moreno@intel.com>