Inst and dev guide versioning: current/latest
These changes implement the versioning scheme for
Installation and Developer Guides for StarlingX.
These guides are release-specific.  Consequently, we
need to make archived, current, and latest versions
of the manuals available as StarlingX releases evolve.

We implemented a two-fold directory structure to keep
the "current" and "latest" versions of the manuals in
the forefront.  "Current" indicates the newest,
officially released version.  "Latest" indicates the
current+1 release (i.e. the release under development).

Aside from "current" and "latest", the list of archived
versions will grow by one with each release.

Changes involved the following:

* Establishing "current" and "latest" directories
  from what existed.
* Updating of intro material and links to be appropriate
  for either "current" or "latest".
* Update of the writer instructions for adding a new
  release of the install and developer manuals.
  material to reflect the folder's new name.
*

Change-Id: I37eee44c31202ac391deebec3348f1e305dbd5ab
Signed-off-by: Scott Rifenbark <srifenbark@gmail.com>