Developer Guide: Tags and note for new mirror.
A note was added to latest/index.rst asking for a new
mirror download due to current issues with installer packages
Also the "2018.10" tag was removed from Download source code
since this has to use master manifest.
From current/index.rst file, the suggestion to use master
manifest was removed due to this is covered in the other
guide.

Depends-on: https://review.openstack.org/#/c/643631

Change-Id: I3cd0e984368a8ed79bb3227978f3fc795dcf8273
Signed-off-by: Hayde Martinez <hayde.martinez.landa@intel.com>