stx-docs refactoring for new doc structure
Changes included updating the StarlingX docs landing
page to conform to a new list of available documents
as follows:

 * StarlingX Introduction
 * Deployment Guides
 * Installation Guides
 * Developer resources
 * Operation Guides
 * Project Specifications
 * REST API Reference
 * Release Notes
 * Contribute

So-called "dead-end" links have an explanation of content
coming.  Any destination associated with an existing story
ID have the link to the story for reference.

The existence of a "current" and "latest" Developer
Guides resulted in the top-level list of doc resources
including both the "Developer Guides" and
"Developer Resources" bullets.  Ideally, the Developer
Guides would be included as part of the "Developer
Resources" area.  For "latest" they are.  However, for
"current" I left them alone since the 2018.10 release
is out there.  This organization can settle out as more
releases launch.

Finally, I did not focus on the writing aspect for this
PR. That part can be cleaned up going forward.

Change-Id: I0442e948d5d9a41add9eb3bdde8882c157a8f752
Signed-off-by: Scott Rifenbark <srifenbark@gmail.com>